Conflicts:
	tests/auto/declarative/qdeclarativedom/data/importlib/sublib/qmldir
	tools/assistant/tools/assistant/assistant.pro
	tools/assistant/tools/assistant/centralwidget.cpp
	tools/assistant/tools/assistant/helpviewer_qtb.cpp
	tools/assistant/tools/assistant/helpviewer_qwv.cpp
	tools/assistant/tools/assistant/mainwindow.cpp

+++ tests/auto/declarative/runall.sh
@@ -41,37 +41,60 @@
 ##
 ############################################################################/
 
-Xnest :7 2>/dev/null &
-sleep 1
-trap "kill $!" EXIT
-export DISPLAY=:7
+if [ "$(uname)" = Linux ]
+then
+    Xnest :7 2>/dev/null &
+    sleep 1
+    trap "kill $!" EXIT
+    export DISPLAY=:7
+    export LANG=en_US
+    kwin 2>/dev/null &
+    sleep 1
+fi
 
-( make -k -j1 install 2>&1;
-  for exe in $(make install | sed -n 's/^install .* "\([^"]*qt4\/tst_[^"]*\)".*/\1/p')
-  do
-     $exe
-  done
-) |
+function filter
+{
+    exe=$1
+    skip=0
     while read line
     do
-        case "$line" in
-        make*Error) echo "$line";;
-        make*Stop) echo "$line";;
-        make*) ;;
-        install*) ;;
-        */qmake*) ;;
-        */bin/moc*) ;;
-        *targ.debug*) ;;
-        g++*) ;;
-        cd*) ;;
-        PASS*) ;;
-        QDEBUG*) ;;
-        Makefile*) ;;
-        Config*) ;;
-        Totals*) ;;
-        \**) ;;
-        ./*) ;;
-        *) echo "$line"
-        esac
+        if [ $skip != 0 ]
+        then
+            let skip=skip-1
+        else
+            case "$line" in
+            make*Error) echo "$line";;
+            make*Stop) echo "$line";;
+            /*/bin/make*) ;;
+            make*) ;;
+            install*) ;;
+            QDeclarativeDebugServer:*Waiting*) ;;
+            QDeclarativeDebugServer:*Connection*) ;;
+            */qmake*) ;;
+            */bin/moc*) ;;
+            *targ.debug*) ;;
+            g++*) ;;
+            cd*) ;;
+            XFAIL*) skip=1;;
+            SKIP*) skip=1;;
+            PASS*) ;;
+            QDEBUG*) ;;
+            Makefile*) ;;
+            Config*) ;;
+            Totals*) ;;
+            \**) ;;
+            ./*) ;;
+            *tst_*) echo "$line" ;;
+            *) echo "$exe: $line"
+            esac
+        fi
     done
+}
+
+make -k -j1 install 2>&1 | filter build
+for exe in $(make install | sed -n 's/^install .* "\([^"]*qt4\/tst_[^"]*\)".*/\1/p')
+do
+    echo $exe
+    $exe 2>&1 | filter $exe
+done
